+++ src/cpu/core_full/load.h
@@ -201,7 +201,7 @@ l_M_Ed:
 		case 0:
 			break;
 		default:
-			LOG(LOG_CPU,LOG_ERROR)("MODRM:Unhandled load %d entry %x",inst.code.extra,inst.entry);
+			LOG(LOG_CPU,LOG_ERROR)("MODRM:Unhandled load %d entry %x",inst.code.extra,(unsigned int)inst.entry);
 			break;
 		}
 		break;
@@ -508,7 +508,7 @@ l_M_Ed:
 		break;
 		}
 	default:
-		LOG(LOG_CPU,LOG_ERROR)("LOAD:Unhandled code %d opcode %X",inst.code.load,inst.entry);
+		LOG(LOG_CPU,LOG_ERROR)("LOAD:Unhandled code %d opcode %X",inst.code.load,(unsigned int)inst.entry);
 		goto illegalopcode;
 }
