+++ src/cpu/core_normal/prefix_none.h
@@ -1165,7 +1165,7 @@
 				else {GetEAa;Push_16(LoadMw(eaa));}
 				break;
 			default:
-				LOG(LOG_CPU,LOG_ERROR)("CPU:GRP5:Illegal Call %2X",which);
+				LOG(LOG_CPU,LOG_ERROR)("CPU:GRP5:Illegal Call %2X",(unsigned int)which);
 				goto illegal_opcode;
 			}
 			break;
